feat: refactor breadcrumb styling
