+++ services/web-app/components/page-breadcrumb/index.js
@@ -4,4 +4,4 @@
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
  */
-export { default } from './breadcrumb'
+export { default } from './page-breadcrumb'
