+++ services/web-app/components/page-header/page-header.module.scss
@@ -26,12 +26,6 @@ $background-header: #d0e2ff;
   }
 }
 
-.breadcrumb {
-  position: absolute;
-  left: 0;
-  width: 100%;
-}
-
 .column {
   position: relative;
   display: flex;
