+++ services/web-app/layouts/layout/layout.module.scss
@@ -11,5 +11,7 @@
 }
 
 .main {
+  // position for PageBreadcrumb
+  position: relative;
   margin: $spacing-09 0 0;
 }
